npm: peerDependencies: +typescript:>=5.5,jest:^29.7 as optional, +eslint:minorVersion constraint
these are critical packages where same major and (to a lesser extent) minimum minor version matter (in the case of optional packages, when installed).
An exception is made for TypeScript which I trust to be backwards compatible even by major version (and if it isn't, it'll break compilation just as much as it would break lint, making the code itself the limiting factor, not the linter).